+++ src/IP.php
@@ -3,8 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\IP;
 
-use Innmind\IP\Exception\DomainException;
-
 /**
  * @psalm-immutable
  */
@@ -24,10 +22,11 @@ abstract class IP
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     final public static function v4(string $address): IPv4
     {
         if (!\filter_var($address, \FILTER_VALIDATE_IP, \FILTER_FLAG_IPV4)) {
-            throw new DomainException($address);
+            throw new \DomainException($address);
         }
 
         /** @psalm-suppress ArgumentTypeCoercion $address cannot be empty here */
@@ -37,16 +36,18 @@ abstract class IP
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     final public static function v6(string $address): IPv6
     {
         if (!\filter_var($address, \FILTER_VALIDATE_IP, \FILTER_FLAG_IPV6)) {
-            throw new DomainException($address);
+            throw new \DomainException($address);
         }
 
         /** @psalm-suppress ArgumentTypeCoercion $address cannot be empty here */
         return new IPv6($address);
     }
 
+    #[\NoDiscard]
     final public function equals(self $self): bool
     {
         return $this->address === $self->address;
@@ -55,6 +56,7 @@ abstract class IP
     /**
      * @return non-empty-string
      */
+    #[\NoDiscard]
     final public function toString(): string
     {
         return $this->address;