+++ src/IPv4.php
@@ -3,8 +3,10 @@ declare(strict_types = 1);
 
 namespace Innmind\IP;
 
-use Innmind\IP\Exception\DomainException;
-use Innmind\Immutable\Maybe;
+use Innmind\Immutable\{
+    Maybe,
+    Attempt,
+};
 
 /**
  * @psalm-immutable
@@ -14,6 +16,7 @@ final class IPv4 extends IP
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     public static function of(string $address): self
     {
         return IP::v4($address);
@@ -22,6 +25,7 @@ final class IPv4 extends IP
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     public static function localhost(): self
     {
         return self::of('127.0.0.1');
@@ -32,13 +36,20 @@ final class IPv4 extends IP
      *
      * @return Maybe<self>
      */
+    #[\NoDiscard]
     public static function maybe(string $address): Maybe
     {
-        try {
-            return Maybe::just(self::of($address));
-        } catch (DomainException $e) {
-            /** @var Maybe<self> */
-            return Maybe::nothing();
-        }
+        return self::attempt($address)->maybe();
+    }
+
+    /**
+     * @psalm-pure
+     *
+     * @return Attempt<self>
+     */
+    #[\NoDiscard]
+    public static function attempt(string $address): Attempt
+    {
+        return Attempt::of(static fn() => self::of($address));
     }
 }